feat: Board 컴포넌트 추가
